+++ platform/nginx-ingress/Install-Nginx-Ingress.ps1
@@ -18,11 +18,11 @@ Write-Output "Release $HelmReleaseName does not exist in namespace $Namespace. I
 
 helm repo add ingress-nginx https://kubernetes.github.io/ingress-nginx
 helm repo update
-helm upgrade $HelmReleaseName --install ingress-nginx `
+helm upgrade --install $HelmReleaseName ingress-nginx `
     --repo https://kubernetes.github.io/ingress-nginx `
     --namespace $Namespace `
     --set controller.service.externalTrafficPolicy=Local `
     --create-namespace
 
 # example call: 
-# .\deploy-ingress-helm.ps1 -HelmReleaseName "event-ingress" -Namespace "event-triangle"
+# .\nginx-ingress\Install-Nginx-Ingress.ps1 -HelmReleaseName "nginx-ingress-dev" -Namespace "ing-dev"