+++ terraform-azdo-libraries/versions.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_providers {
+    azuredevops = {
+      source  = "microsoft/azuredevops"
+      version = ">=1.1.1"
+    }
+  }
+
+  backend "azurerm" {}
+}