+++ CHANGELOG.md
@@ -46,3 +46,5 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Platform: Change PostgeSQL to be cluster IP
 - Disable terraform log info
 - Add AKS node pool name
+- Platform updates
+- Azure DevOps library terraform